+++ client/python/gradio_client/utils.py
@@ -25,7 +25,6 @@ from typing import (
     Any,
     Literal,
     NewType,
-    Optional,
     TypedDict,
     Union,
     get_args,
@@ -192,11 +191,11 @@ class Status(Enum):
 
 @dataclass
 class ProgressUnit:
-    index: Optional[int]
-    length: Optional[int]
-    unit: Optional[str]
-    progress: Optional[float]
-    desc: Optional[str]
+    index: int | None
+    length: int | None
+    unit: str | None
+    progress: float | None
+    desc: str | None
 
     @classmethod
     def from_msg(cls, data: list[dict]) -> list[ProgressUnit]: