+++ demo/native_plots/run.py
@@ -1,8 +1,8 @@
 import gradio as gr
 
-from scatter_plot_demo import scatter_plots
-from line_plot_demo import line_plots
-from bar_plot_demo import bar_plots
+from scatter_plot_demo import scatter_plots  # type: ignore
+from line_plot_demo import line_plots  # type: ignore
+from bar_plot_demo import bar_plots  # type: ignore
 
 with gr.Blocks() as demo:
     with gr.Tabs():