+++ demo/text_analysis/run.py
@@ -1,8 +1,8 @@
 import gradio as gr
 import os
 os.system('python -m spacy download en_core_web_sm')
-import spacy
-from spacy import displacy
+import spacy  # type: ignore
+from spacy import displacy  # type: ignore
 
 nlp = spacy.load("en_core_web_sm")