+++ demo/yolov10_webcam_stream/inference.py
@@ -1,5 +1,5 @@
 import time
-import cv2
+import cv2  # type: ignore
 import numpy as np
 import onnxruntime  # type: ignore