+++ gradio/components/upload_button.py
@@ -204,7 +204,7 @@ class UploadButton(Component):
             return downloaded_files
         if client_utils.is_http_url_like(value):
             downloaded_file = processing_utils.save_url_to_cache(
-                value, self.GRADIO_CACHE
+                str(value), self.GRADIO_CACHE
             )
             return downloaded_file
         else: