+++ js/form/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gradio/form
 
+## 0.2.23
+
+### Dependency updates
+
+- @gradio/atoms@0.16.5
+- @gradio/icons@0.13.0
+
 ## 0.2.22
 
 ### Dependency updates