+++ js/nativeplot/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gradio/nativeplot
 
+## 0.7.4
+
+### Dependency updates
+
+- @gradio/statustracker@0.10.17
+- @gradio/atoms@0.16.5
+- @gradio/icons@0.13.0
+
 ## 0.7.3
 
 ### Dependency updates