+++ js/slider/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gradio/slider
 
+## 0.6.16
+
+### Dependency updates
+
+- @gradio/statustracker@0.10.17
+- @gradio/atoms@0.16.5
+
 ## 0.6.15
 
 ### Dependency updates