+++ js/textbox/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gradio/textbox
 
+## 0.10.20
+
+### Dependency updates
+
+- @gradio/statustracker@0.10.17
+- @gradio/atoms@0.16.5
+- @gradio/icons@0.13.0
+
 ## 0.10.19
 
 ### Dependency updates