+++ js/tootils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @self/tootils
 
+## 0.7.28
+
+### Dependency updates
+
+- @gradio/statustracker@0.10.17
+
 ## 0.7.27
 
 ### Dependency updates